customers/serializers.py
from rest_framework import serializers
from customers.models import Customer


class CustomerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Customer
        fields = ('id', 'first_name', 'last_name', 'email', 'gender', 'company', 'city', 'title', 'latitude', 'longitude')

