requirements.txt
Django==3.0
djangorestframework==3.11.0
psycopg2-binary==2.8.5
django-extensions==3.0.5
geocoder==1.38.1
drf-yasg==1.17.1